Fix error_prone violations

+++ log-manager/src/test/java/io/airlift/log/TestLogHistoryManager.java
@@ -75,7 +75,7 @@ public class TestLogHistoryManager
 
     private static void assertLogFiles(LogHistoryManager logHistoryManager, List<LogFileName> expected)
     {
-        assertThat(logHistoryManager.getTotalSize()).isEqualTo(expected.size() * FILE_SIZE);
+        assertThat(logHistoryManager.getTotalSize()).isEqualTo((long) expected.size() * FILE_SIZE);
         List<LogFileName> files = logHistoryManager.getFiles().stream()
                 .sorted()
                 .collect(toImmutableList());
